+++ features/greeting.js
@@ -1,28 +1,3 @@
-const randomArrayItem = array => array[Math.floor(Math.random() * array.length)];
+const greetingController = require('./greeting/index');
 
-const queries = [
-  'Hello',
-  'Good morning',
-  'G\'day',
-  '\^Hi\$',
-  'Morning',
-  'Hey',
-];
-
-const replies = [
-  'Hey!',
-  'Morning, :smiley:',
-  'Nice... :smiley:',
-  'Hello! :smiley:',
-  'Hi, how are you?',
-];
-
-module.exports = (controller) => {
-  controller.hears(
-    queries,
-    ['message', 'direct_message'],
-    async (bot, message) => {
-      await bot.reply(message, { text: randomArrayItem(replies) });
-    },
-  );
-};
+module.exports = greetingController;
