+++ features/greeting/queries.js
@@ -1,0 +1,8 @@
+module.exports = [
+  'Hello',
+  'Good morning',
+  'G\'day',
+  '\^Hi\$',
+  'Morning',
+  'Hey',
+];
