+++ features/greeting/replies.js
@@ -1,0 +1,7 @@
+module.exports = [
+  'Hey!',
+  'Morning, :smiley:',
+  'Nice... :smiley:',
+  'Hello! :smiley:',
+  'Hi, how are you?',
+];
